Do not display "add" option when the value already exists in options.
Previous behavior would attempt to show both the "add <input value" option to
add the new value and the value from the autocomplete options. The code would
give them both the same key, so react would dedupe them and the latter value
was lost. This simplifies things and only shows the autocomplete value it this
case.